+++ posthog-android/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - add evaluation tags to android SDK ([#301](https://github.com/PostHog/posthog-android/pull/301))
 - feat: add manual captureException ([#300](https://github.com/PostHog/posthog-android/issues/300))
+- feat: add exception autocapture ([#305](https://github.com/PostHog/posthog-android/issues/305))
 
 ## 3.23.0 - 2025-10-06